requirements.txt
# Minimal requirements for Auto Certificate Generator
# Generated by analyzing actual imports in the codebase (2025-11-05)

# Flask web framework and security
Flask>=3.0.0
flask-talisman>=1.0.0
flask-wtf>=1.2.0
flask-limiter>=3.5.0

# Google APIs (Drive, Sheets, Firebase)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
firebase-admin>=6.0.0
gspread>=5.11.0

# Image processing (PIL/Pillow)
Pillow>=10.0.0

# PDF processing
reportlab>=4.0.0
PyPDF2>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: MySQL support (only if using MySQL instead of SQLite)
# Uncomment if you need MySQL:
# pymysql>=1.1.0
